Further improvement

+++ src/cli/subcommand/install.rs
@@ -78,7 +78,11 @@ impl CommandExecute for Install {
                 let install_plan_string = tokio::fs::read_to_string(&RECEIPT_LOCATION)
                     .await
                     .wrap_err("Reading plan")?;
-                Some(serde_json::from_str(&install_plan_string)?)
+                Some(
+                    serde_json::from_str(&install_plan_string).wrap_err_with(|| {
+                        format!("Unable to parse existing receipt `{RECEIPT_LOCATION}`, it may be from an incompatible version of `nix-installer`. Try running `/nix/nix-installer uninstall`, then installing again.")
+                    })?,
+                )
             },
             false => None,
         };
@@ -97,11 +101,11 @@ impl CommandExecute for Install {
                         if let Err(e) = existing_receipt.check_compatible() {
                             eprintln!(
                                 "{}", 
-                                format!("
+                                format!("\
                                     {e}\n\
                                     \n\
                                     Found existing plan in `{RECEIPT_LOCATION}` which was created by a version incompatible `nix-installer`.\n\
-                                    If you are trying to upgrade Nix, try running `sudo -i nix upgrade-nix` instead.
+                                    If you are trying to upgrade Nix, try running `sudo -i nix upgrade-nix` instead.\n\
                                     If you are trying to install Nix over an existing install (from an incompatible `nix-installer` install), try running `/nix/nix-installer uninstall` then try to install again.\n\
                                     If you are using `nix-installer` in an automated curing process and seeing this message, consider pinning the version you use via https://github.com/DeterminateSystems/nix-installer#accessing-other-versions.\n\
                                 ").red()
@@ -147,6 +151,20 @@ impl CommandExecute for Install {
 
                 match existing_receipt {
                     Some(existing_receipt) => {
+                        if let Err(e) = existing_receipt.check_compatible() {
+                            eprintln!(
+                                "{}", 
+                                format!("\
+                                    {e}\n\
+                                    \n\
+                                    Found existing plan in `{RECEIPT_LOCATION}` which was created by a version incompatible `nix-installer`.\n\
+                                    If you are trying to upgrade Nix, try running `sudo -i nix upgrade-nix` instead.\n\
+                                    If you are trying to install Nix over an existing install (from an incompatible `nix-installer` install), try running `/nix/nix-installer uninstall` then try to install again.\n\
+                                    If you are using `nix-installer` in an automated curing process and seeing this message, consider pinning the version you use via https://github.com/DeterminateSystems/nix-installer#accessing-other-versions.\n\
+                                ").red()
+                            );
+                            return Ok(ExitCode::FAILURE)
+                        }
                         if existing_receipt.planner.typetag_name() != builtin_planner.typetag_name() {
                             eprintln!("{}", format!("Found existing plan in `{RECEIPT_LOCATION}` which used a different planner, try uninstalling the existing install with `{uninstall_command}`").red());
                             return Ok(ExitCode::FAILURE)
